% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvm.test.R
\name{cvm.test}
\alias{cvm.test}
\title{The two-sample Cramer-von Mises test using R}
\usage{
cvm.test(x, y)
}
\arguments{
\item{x}{The sample from distribution X}

\item{y}{The sample from distribution Y}
}
\value{
score of the two-sample Cramer-von Mises test for given samples and p value
}
\description{
The two-sample Cramer-von Mises test using R
}
\examples{
\dontrun{
setseed(1121)
x <-  rpois(30,2)
y <-  rpois(30,3)
cvm.test(x,y)
}
}
