% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rayleigh.MC.R
\name{Rayleigh.MC}
\alias{Rayleigh.MC}
\title{Generate samples from Rayleigh distribution}
\usage{
Rayleigh.MC(x, sigma, m = 10000, antithetic = TRUE)
}
\arguments{
\item{x}{vector of quantiles}

\item{sigma}{the parameter of Rayleigh density}

\item{m}{the sample size}

\item{antithetic}{using antithetic method or not}
}
\value{
a random sample of size \code{length(x)}
}
\description{
Generate samples from Rayleigh distribution
}
\examples{
\dontrun{
X <- sample.chain(100,1)
print(Gelman.Rubin(X))
}
}
