% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcauchy.MH.R
\name{rcauchy.MH}
\alias{rcauchy.MH}
\title{Generate random samples from Cauchy distribution by MH sampler}
\usage{
rcauchy.MH(n, theta = 0, eta = 1)
}
\arguments{
\item{n}{number of observations.}

\item{theta}{location parameter}

\item{eta}{scale parameter}
}
\value{
a random sample of size \code{n}
}
\description{
Generate random samples from Cauchy distribution by MH sampler
}
\examples{
\dontrun{
rcauchy(100)
}
}
