% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.chain.R
\name{sample.chain}
\alias{sample.chain}
\title{Gnerate Metropolis chain from giving multinomial distribution}
\usage{
sample.chain(N, start)
}
\arguments{
\item{N}{the length of chain}

\item{start}{the starting value}
}
\value{
a random sample of size \code{N}
}
\description{
Gnerate Metropolis chain  from giving multinomial distribution
}
\examples{
\dontrun{
rnR <- sample.chain(100,1)
plot(rnR,type='l')
}
}
