\name{StatComp18087-package}
\alias{StatComp18087-package}
\docType{package}
\title{
  This package provides Several R function are included in this package for illustration.
}
\description{
  R and Rcpp versions of gibbs sampler and vaac function are provided in this package (\code{cvm.test} and \code{Rayleigh.MC} and \code{rcauchy.MH})
}
\author{
18087

Maintainer: 18087 Developer <731754338@qq.com>
}
\keyword{ package }
